show recommentdation

+++ lib/admin page&staff/AdminModel/product_model.dart
@@ -7,6 +7,7 @@
 
 import 'dart:convert';
 
+
 List<ProductModel> productModelFromJson(String str) => List<ProductModel>.from(json.decode(str).map((x) => ProductModel.fromJson(x)));
 
 String productModelToJson(List<ProductModel> data) => json.encode(List<dynamic>.from(data.map((x) => x.toJson())));
@@ -54,12 +55,12 @@ class ProductModel {
 class Item {
     Item({
         required this.itemImage,
-        required this.price,
+         required this.price,
         required this.name,
     });
 
     String itemImage;
-    int price;
+    double price;
     String name;
 
     factory Item.fromJson(Map<String, dynamic> json) => Item(
